SNOW-1271212 unit tests for CSV Data Conversion
